+++ photutils/utils/tests/test_colormaps.py
@@ -15,4 +15,5 @@ def test_colormap():
     ncolors = 100
     cmap = make_random_cmap(ncolors, seed=0)
     assert len(cmap.colors) == ncolors
-    assert_allclose(cmap.colors[0], [0.36951484, 0.42125961, 0.65984082])
+    assert cmap.colors.shape == (100, 4)
+    assert_allclose(cmap.colors[0], [0.36951484, 0.42125961, 0.65984082, 1.])